[Refactor] 가독성 향상을 위한 코드 정리 및 파일로 분리
