+++ lib/presentation/pages/search/components/search_bar.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+
+class SearchBar extends StatelessWidget {
+  const SearchBar({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: double.infinity,
+      padding: const EdgeInsets.all(8),
+      decoration: BoxDecoration(
+        color: Colors.white.withOpacity(0.5),
+        borderRadius: BorderRadius.circular(8),
+        border: Border.all(color: Colors.grey),
+      ),
+      child: Row(
+        children: [
+          SvgPicture.asset(
+            'assets/icons/search.svg',
+            colorFilter: ColorFilter.mode(Colors.black, BlendMode.srcIn),
+            width: 24,
+            height: 24,
+          ),
+          const SizedBox(width: 16),
+          Text(
+            'sdfsdfsd',
+            style: TextStyle(
+              fontSize: 16,
+              color: Colors.black,
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
