+++ lib/presentation/pages/search/search_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-import 'package:weather_app_flutter/core/theme/constant/app_colors.dart';
+
+import 'components/search_results.dart';
 
 class SearchPage extends StatelessWidget {
   const SearchPage({super.key});
@@ -10,82 +10,10 @@ class SearchPage extends StatelessWidget {
     return Column(
       children: [
         SearchBar(),
-        SizedBox(height: 20),
+        const SizedBox(height: 20),
         SearchResults(),
       ],
     );
   }
 }
 
-class SearchBar extends StatelessWidget {
-  const SearchBar({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: double.infinity,
-      padding: const EdgeInsets.all(8),
-      decoration: BoxDecoration(
-        color: Colors.white.withOpacity(0.5),
-        borderRadius: BorderRadius.circular(8),
-        border: Border.all(color: Colors.grey),
-      ),
-      child: Row(
-        children: [
-          SvgPicture.asset(
-            'assets/icons/search.svg',
-            colorFilter: ColorFilter.mode(Colors.black, BlendMode.srcIn),
-            width: 24,
-            height: 24,
-          ),
-          const SizedBox(width: 16),
-          Text(
-            'sdfsdfsd',
-            style: TextStyle(
-              fontSize: 16,
-              color: Colors.black,
-            ),
-          )
-        ],
-      ),
-    );
-  }
-}
-
-
-class SearchResults extends StatelessWidget {
-  const SearchResults({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Expanded(
-      child: ListView.separated(
-        separatorBuilder: (context, index) => const Divider(),
-        itemCount: 12,
-        itemBuilder: (context, index) => Padding(
-          padding: const EdgeInsets.symmetric(vertical: 4),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                'Republic Of Korea',
-                style: TextStyle(
-                  fontSize: 20,
-                  fontWeight: FontWeight.bold,
-                  color: AppColors.textColor,
-                ),
-              ),
-              Text(
-                'KR',
-                style: TextStyle(
-                  fontSize: 16,
-                  color: AppColors.textColor,
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
